Update requests library due to vulnerability
Github alerted me on this:

The Requests package through 2.19.1 before 2018-09-14 for Python sends
an HTTP Authorization header to an http URI upon receiving a
same-hostname https-to-http redirect, which makes it easier for remote
attackers to discover credentials by sniffing the network.
https://nvd.nist.gov/vuln/detail/CVE-2018-18074

We don't deal with authorization here, but it doesn't hurt to get this
dependency updated.